extract mobile layout
